// lib/generated/prisma/models/Company.ts
export type CompanyModel = runtime.Types.Result.DefaultSelection<Prisma.$CompanyPayload>
// ...
export type $CompanyPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Company"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    slug: string
    name: string | null
    logo: string | null
    description: string | null
    developed: number[]
    published: number[]
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["company"]>
  composites: {}
}